+++ packages/apps-config/src/api/spec/index.ts
@@ -4,6 +4,7 @@
 import type { OverrideBundleDefinition } from '@polkadot/types/types';
 
 import acala from './acala';
+import apron from './apron';
 import aresParachain from './ares-parachain';
 import bifrost from './bifrost';
 import bifrostParachain from './bifrost-parachain';
@@ -22,11 +23,13 @@ import datahighwayParachain from './datahighway';
 import dock from './dock';
 import dotmog from './dotmog';
 import dusty from './dusty';
+import eave from './eave';
 import edgeware from './edgeware';
 import encointerNodeNotee from './encointer-node-notee';
 import encointerNodeTeeproxy from './encointer-node-teeproxy';
 import encointerPara from './encointer-para';
 import equilibrium from './equilibrium';
+import galital from './galital';
 import galois from './galois';
 import hanonycash from './hanonycash';
 import hydrate from './hydrate';
@@ -38,8 +41,12 @@ import kilt from './kilt';
 import kulupu from './kulupu';
 import laminar from './laminar';
 import litentry from './litentry';
+import manta from './manta';
 import moonbeam from './moonbeam';
+import mybank from './mybank';
+import nftmart from './nftmart';
 import nodle from './nodle';
+import parami from './parami';
 import phala from './phala';
 import phalaParachain from './phala-parachain';
 import phoenix from './phoenix';
@@ -57,9 +64,11 @@ import subdao from './subdao';
 import subsocial from './subsocial';
 import ternoa from './ternoa';
 import trustbase from './trustbase';
-import uniarts from './uniarts';
+import uart from './uart';
 import unique from './unique';
+import unitv from './unitv';
 import vln from './vln';
+import zeitgeist from './zeitgeist';
 import zenlink from './zenlink';
 import zero from './zero';
 
@@ -72,6 +81,7 @@ const spec: Record<string, OverrideBundleDefinition> = {
   Equilibrium: equilibrium,
   VLN: vln,
   acala,
+  apron,
   'ares-parachain': aresParachain,
   bifrost: bifrost,
   'bifrost-parachain': bifrostParachain,
@@ -85,8 +95,10 @@ const spec: Record<string, OverrideBundleDefinition> = {
   'clover-rococo': cloverRococo,
   crust,
   'crust-parachain': crust,
+  'cumulus-subsocial-parachain': subsocial,
   'cumulus-test-parachain': testPara,
   'datahighway-parachain': datahighwayParachain,
+  dawn: eave,
   'dev-parachain': zenlink,
   'dock-main-runtime': dock,
   'dock-testnet': dock,
@@ -96,6 +108,7 @@ const spec: Record<string, OverrideBundleDefinition> = {
   'encointer-node-notee': encointerNodeNotee,
   'encointer-node-teeproxy': encointerNodeTeeproxy,
   'encointer-parachain': encointerPara,
+  'galital-collator': galital,
   'hack-hydra-dx': hydrate,
   hanonycash,
   'hydra-dx': hydrate,
@@ -108,15 +121,19 @@ const spec: Record<string, OverrideBundleDefinition> = {
   laminar,
   litentry,
   mandala: acala,
+  manta,
   'mashnet-node': kilt,
   'mathchain-galois': galois,
   'moonbase-alphanet': moonbeam,
   moonbeam,
   'moonbeam-standalone': moonbeam,
+  'mybank.network Testnet': mybank,
   nft: unique,
+  nftmart,
   'node-moonbeam': moonbeam,
   'node-polkadex': polkadex,
   'nodle-chain': nodle,
+  parami,
   'phala-collator': phalaParachain,
   'phala-node': phala,
   'phoenix-node': phoenix,
@@ -129,12 +146,16 @@ const spec: Record<string, OverrideBundleDefinition> = {
   snowbridge,
   'sora-substrate': soraSubstrate,
   stafi,
+  steam: eave,
   subdao,
   subsocial,
   subzero: zero,
   ternoa,
   trustbase,
-  uniarts
+  uart,
+  'unit-node': unitv,
+  'unit-parachain': unitv,
+  zeitgeist: zeitgeist
 };
 
 export default spec;