+++ packages/apps-config/src/endpoints/testing.ts
@@ -271,10 +271,39 @@ export function createTesting (t: TFunction): LinkOption[] {
       }
     },
     {
+      info: 'zeitgeist',
+      text: t('rpc.zeitgeist', 'Zeitgeist Battery Park', { ns: 'apps-config' }),
+      providers: {
+        Zeitgeist: 'wss://bp-rpc.zeitgeist.pm'
+      }
+    },
+    {
       info: 'zero',
       text: t('rpc.zero', 'Zero', { ns: 'apps-config' }),
       providers: {
         ZERO: 'wss://alphaville.zero.io'
+      }
+    },
+    {
+      info: 'mybank',
+      text: t('rpc.mybank', 'mybank.network', { ns: 'apps-config' }),
+      providers: {
+        MYBANK: 'wss://app.mybank.network/substrate'
+      }
+    },
+    {
+      info: 'unitv',
+      text: t('rpc.unitv', 'Unit Network', { ns: 'apps-config' }),
+      providers: {
+        'Unit Network': 'wss://unitventures.io/'
+      }
+    },
+    {
+      info: 'nftmart',
+      text: t('rpc.nftmart', 'Nftmart', { ns: 'apps-config' }),
+      providers: {
+        NftmartDev: 'ws://8.136.111.191:9944',
+        NftmartStaging: 'ws://82.157.37.77:9944'
       }
     }
   ]);